Fill in the client side APIs and related IPC.
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@49694 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lwtask/include/lwtaskdef.h
@@ -72,9 +72,9 @@ typedef DWORD LW_TASK_ARG_FLAG;
 
 typedef struct _LW_TASK_STATUS
 {
-    DWORD  dwError;
+    DWORD dwError;
 
-    double percentComplete;
+    DWORD percentComplete;
 
 } LW_TASK_STATUS, *PLW_TASK_STATUS;
